Move executor implementation to separate file
